back/auto-feed/requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
fastapi==0.115.4
firebase_admin==6.5.0
langchain==0.3.7
langchain_community==0.3.5
openai==0.28.0
protobuf==5.28.3
pydantic==2.9.2
python-dotenv==1.0.1
pytrends==4.9.2
Requests==2.32.3
selenium==4.26.1
transformers==4.45.2
urllib3==2.2.3
uvicorn==0.32.0
webdriver_manager==4.0.2
